+++ pair.h
@@ -1,11 +1,13 @@
 #ifndef PAIR_H
 #define PAIR_H
 
+#include <inttypes.h>
+
 typedef struct pair Pair;
 
-Pair *newpair(int key, char *value);
+Pair *newpair(uintmax_t key, char *value);
 void freepair(Pair *pair);
-int getkey(Pair *pair);
+uintmax_t getkey(Pair *pair);
 char *getvalue(Pair *pair);
 void setvalue(Pair *pair, char *value);
